refactor(admin-console): extract CreateIdpModal from IdentityProviders.tsx
IdentityProviders.tsx held both the IdP list page and a self-contained
CreateIdpModal (SAML registration form), pulling 17 imports — over the
high-coupling threshold of 16. The modal owned the only Modal / FormField /
Input / Textarea uses, so extracting it to CreateIdpModal.tsx drops
IdentityProviders.tsx to 15 imports (off the high-coupling list).

This also sets up the cross-app dedup: application-admin-console has a
near-identical IdentityProviders page + modal; once both modals are isolated
the shared extraction into a package becomes a mechanical follow-up.

Relates #1418.

## Regression analysis
- Pure structural move: CreateIdpModal's body + props are byte-identical, only
  relocated and imported back. admin-console typecheck passes; all 143 tests
  pass unchanged. describeIdpError stays used by the page's load/delete paths.

## Physical impact
- NO-OP on CFn / deployed artifacts. Frontend source reorganization only.
